requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argcomplete=1.12.3=pypi_0
asgiref=3.4.1=pypi_0
audioread=2.1.9=py37h03978a9_1
backcall=0.2.0=pypi_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
certifi=2021.10.8=py37haa95532_0
cffi=1.15.0=py37h2bbff1b_0
charset-normalizer=2.0.7=pypi_0
colorama=0.4.4=pypi_0
cryptography=36.0.0=py37h21b164f_0
cudatoolkit=11.2.2=h933977f_9
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=pypi_0
decorator=4.4.0=py37_1
django=3.2.9=pypi_0
encoder=1.1=pypi_0
entrypoints=0.3=pypi_0
ffmpeg=4.2.2=he774522_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
futures=2.2.0=pypi_0
gevent=1.4.0=pypi_0
goslate=1.5.1=pypi_0
greenlet=0.4.15=pypi_0
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imageio=2.9.0=pyhd3eb1b0_0
imageio-ffmpeg=0.4.5=pyhd8ed1ab_0
importlib-metadata=4.8.1=pypi_0
inflect=5.3.0=py37haa95532_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.2=pypi_0
ipython=7.29.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
jupyter-client=7.0.6=pypi_0
jupyter-core=4.9.1=pypi_0
kiwisolver=1.3.1=py37hd77b12b_0
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libflac=1.3.3=h0e60522_1
liblapack=3.8.0=20_mkl
libogg=1.3.5=h2bbff1b_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.37=h2a8f88b_0
librosa=0.7.0=py_0
libsndfile=1.0.31=h0e60522_1
libtiff=4.2.0=hd0e1b90_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.0=h2bbff1b_0
llvmlite=0.31.0=py37ha925a31_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.4.3=py37haa95532_0
matplotlib-base=3.4.3=py37h49ac443_0
matplotlib-inline=0.1.3=pypi_0
mkl=2020.2=256
moviepy=1.0.1=py_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nest-asyncio=1.5.1=pypi_0
numba=0.48.0=py37h47e9c7a_0
numpy=1.17.1=py37hc71023c_0
olefile=0.46=py37_0
opencv-contrib-python=4.5.4.60=pypi_0
opencv-python=4.1.0.25=pypi_0
openssl=1.0.1k=1
parso=0.8.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.4=py37haa95532_0
portaudio=19.6.0=he774522_4
proglog=0.1.9=py_0
prompt-toolkit=3.0.21=pypi_0
pyaudio=0.2.11=py37hfa6e2cd_1
pycparser=2.19=pypi_0
pygments=2.10.0=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.6.0=py37ha878b3d_6
pysocks=1.7.1=py37_1
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.7.0=hea74fb7_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytz=2021.3=pypi_0
pywin32=302=pypi_0
pyzmq=22.3.0=pypi_0
qt=5.6.2=vc14h6f8c307_12
requests=2.26.0=pyhd3eb1b0_0
resampy=0.2.2=py_0
scikit-learn=1.0.1=py37hf11a4ad_0
scipy=1.5.3=py37h6db1a17_0
setuptools=58.0.4=py37haa95532_0
sip=4.18.1=py37hd77b12b_2
six=1.12.0=pypi_0
sounddevice=0.4.3=pypi_0
sqlite=3.36.0=h2bbff1b_0
sqlparse=0.4.2=pypi_0
tbb=2021.4.0=h59b6b97_0
tencentcloud-sdk-python=3.0.541=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
torch=1.10.0+cu113=pypi_0
torchaudio=0.10.0+cu113=pypi_0
torchvision=0.11.1+cu113=pypi_0
tornado=6.1=py37h2bbff1b_0
tqdm=4.45.0=pypi_0
traitlets=5.1.1=pypi_0
typing-extensions=4.0.1=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.7=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wave=0.0.2=pypi_0
wcwidth=0.2.5=pypi_0
websocket=0.2.1=pypi_0
websocket-client=0.56.0=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
xz=5.2.5=h62dcd97_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
